+++ plugins/GSdx/Renderers/DX11/GSTextureCache11.h
@@ -21,7 +21,7 @@
 
 #pragma once
 
-#include "Renderers/Common/GSTextureCache.h"
+#include "Renderers/HW/GSTextureCache.h"
 #include "GSDevice11.h"
 
 class GSTextureCache11 : public GSTextureCache
