+++ plugins/GSdx/Renderers/HW/GSTextureCache.h
@@ -21,9 +21,9 @@
 
 #pragma once
 
-#include "GSRenderer.h"
-#include "GSFastList.h"
-#include "GSDirtyRect.h"
+#include "Renderers/Common/GSRenderer.h"
+#include "Renderers/Common/GSFastList.h"
+#include "Renderers/Common/GSDirtyRect.h"
 
 class GSTextureCache
 {
